fyle/requirements.txt
django==1.9.2
django-extensions==1.6.1
django-rest-swagger==0.3.5
requests==2.11.1
djangorestframework==3.3.2
django-user-roles
psycopg2


